feat: Rendering for two players

+++ star-dash/star-dash/Rendering/Utils/PlayerViewLayout.swift
@@ -2,5 +2,5 @@ import UIKit
 
 struct PlayerViewLayout {
     let superview: UIView
-    let rotation: Float
+    let rotation: CGFloat
 }
